requirements.txt
Flask
Flask-Login
Flask-WTF
wtforms
werkzeug
requests
python-dotenv